code cleanup and static office dataset
